Freighter C!
This should hopefully add the Freighter C and no other bugs or problems most certainly hopefully maybe idk.

Fed Freighter alignments are most likely needed to make them all more centre. This one in particular may be off thanks to Superluminal 2.
Also not sure if the Engine effects are working on this due to SL, and furthermore I am not sure if the floor image is working. It is meant to share one with the type A as I am copying vanilla in that multiuse of floor images.